// _archive/dstream-work-20260301-200950/packages/protocol/src/types.ts
export type NostrTag = string[];

export interface NostrEvent {
  id?: string;
  pubkey: string;
  created_at: number;
  kind: number;
  tags: NostrTag[];
  content: string;
  sig?: string;
}

export const NOSTR_KINDS = {
  STREAM_ANNOUNCE: 30311,
  STREAM_CHAT: 1311,
  PRESENCE: 30312,
  P2P_SIGNAL: 8108,
  MANIFEST_ROOT: 30313,
  XMR_RECEIPT: 30314,
  GUILD: 30315,
  P2P_BYTES_RECEIPT: 30316,
  STREAM_MOD_ACTION: 30317,
  STREAM_MOD_ROLE: 30318,
  GUILD_MEMBERSHIP: 30319,
  GUILD_ROLE: 30320,
  APP_DISCOVERY_MOD: 30321
} as const;
// ...
export interface GuildFeaturedStreamRef {
  streamPubkey: string;
  streamId: string;
}

export interface Guild {
  pubkey: string;
  guildId: string;
  name: string;
  about?: string;
  image?: string;
  topics: string[];
  featuredStreams: GuildFeaturedStreamRef[];
  createdAt: number;
  raw: NostrEvent;
}